requirements.txt
pydub==0.25.1
google-cloud-speech==2.21.0
python-dotenv==1.0.0 